telegramm export/requirements.txt
telethon>=1.24.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio